Fix float datatype issue, minor names fixes, add different files multithreaded test

+++ HDF5CSharpWrapper.ConsoleApp/Program.cs
@@ -25,8 +25,8 @@
             /*int[] intArray = { 1, 2, 3, 4, 5 };
             int[,] intTwoDimArray = { { 1, 2 },
                                       { 3, 4 } };
-            double[] doubleArray = { 25.0 };
-            double[,] doubleTwoDimArray = { { 1.1, 2.2 },
+            float[] doubleArray = { 25.0 };
+            float[,] doubleTwoDimArray = { { 1.1, 2.2 },
                                             { 3.3, 4.4} };
             byte[] byteArray = { 1, 0, 1 };
             byte[,] byteTwoDimArray = { { 1, 0 },
@@ -41,15 +41,15 @@
 
             var dataset = datasets.SetDataset<int>(file, "int", intArray);
             var dataset2 = datasets.SetDataset<int>(file, "intTwoDim", intTwoDimArray);
-            var dataset3 = datasets.SetDataset<double>(file, "double", doubleArray);
-            var dataset4 = datasets.SetDataset<double>(file, "doubleTwoDim", doubleTwoDimArray);
+            var dataset3 = datasets.SetDataset<float>(file, "double", doubleArray);
+            var dataset4 = datasets.SetDataset<float>(file, "doubleTwoDim", doubleTwoDimArray);
             var dataset5 = datasets.SetDataset<byte>(file, "byte", byteArray);
             var dataset6 = datasets.SetDataset<byte>(file, "byteTwoDim", byteTwoDimArray);
             var dataset7 = datasets.SetDataset<byte>(file, "charTwoDim", charArray);
             var dataset8 = datasets.SetDataset<sbyte>(file, "imageColor", imageColorArray);
 
-            var dset = datasets.GetDataset<Double>(file2, "float");
-            var dset2 = datasets.GetDataset<Double>(file2, "/Arrays/float");
+            var dset = datasets.GetDataset<float>(file2, "float");
+            var dset2 = datasets.GetDataset<float>(file2, "/Arrays/float");
             var dset3 = datasets.GetDataset<Int32>(file2, "/int");
             var dset4 = datasets.GetDataset<Int32>(file2, "/Arrays/int");
             var dset5 = datasets.GetDataset<SByte>(file2, "/color_image");
